creation of 2 dataframes
- 'gdp' dataframe has been created and preprocessed. It only contains the countries that are included in the who covid dataset and the gdp dataset.
- 'Qcovid' dataframe contains the number of cases per quarter for each country that are contained in both datasets.

starting cases variable is the starting number of cases per quarter. (ie Q1 starts 1-1-2020.)

ending cases varibale is the number of cases at the end of the quarter(ie. Q1 ends 3-31-2020.)

total cases variable is the number of new cases at that partcular quarter. (aka the difference of ending cases and starting cases)

same rational goes for the death variables